requirements.txt
streamlit==1.30.0
pandas==2.1.0
numpy==1.25.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.18.0
scikit-learn==1.3.0
transformers==4.34.0  # Optional, only if you want advanced translation
torch==2.0.1  # Optional, only if you want advanced translation
PyPDF2==3.0.1
python-docx==0.8.11
#textract==1.6.5  # Removed due to compatibility issues
pdfplumber==0.10.2
langdetect==1.0.9
googletrans==4.0.0-rc1
# indictrans==1.0.3  # Changed to optional
spacy==3.6.1
nltk==3.8.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
boto3==1.28.38
sentence-transformers==2.2.2  # Removed per user request
tqdm==4.66.1